pyproject.toml
[tool.poetry]
name = "drf-react-todo"
version = "0.1.0"
description = ""
authors = ["vyacheslav.v <vyacheslav.v@itfox-web.com>"]
readme = "README.md"
packages = [{include = "drf_react_todo"}]

[tool.poetry.dependencies]
python = "^3.10"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
django = "^4.1.3"
envparse = "^0.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
